Added universe files for DCOS package for PR 1
Signed-off-by: petter <petterhg@kth.se>

This commit, followed by previous ones creates a DC/OS package for Scylla. The package and all its dependencies can be found under /dist/dcos in the project hierarchy.
This package adds features to the original Docker image as well as lets user interactively set their ScyllaDB cluster by using the DC/OS install GUI.
Users can now freely create an arbitrary large cluster of nodes, with any level of security, utilizing most (relevant) features of Scylla via the scylla.yaml file etc.
In /containers, the modified Docker image is located with all auxiliary scripts.
In /universe, all config files exists that are necassary for the ScyllaDB DC/OS package.
Note that /universe will also be committed separately to the DC/OS universe package for it to be supplied to the community.
It is therefore necessary to first establish and agree upon what features this package should contain within the scylla community before I commit this to the DC/OS project.

By making a contribution to this project, I certify that:
The contribution was created in whole or in part by me and I have the right to submit it under the open source license indicated in the file; or

The contribution is based upon previous work that, to the best of my knowledge, is covered under an appropriate open source license and I have the right under that license to submit that work with modifications, whether created in whole or in part by me, under the same open source license (unless I am permitted to submit under a different license), as indicated in the file; or

The contribution was provided directly to me by some other person who certified (a), (b) or (c) and I have not modified it.

I understand and agree that this project and the contribution are public and that a record of the contribution (including all personal information I submit with it, including my sign-off) is maintained indefinitely and may be redistributed consistent with this project or the open source license(s) involved.